refactor: move the shared terrain sampling into terrain.ts
